+++ crc16_arc.h
@@ -4,10 +4,10 @@
  * @param byte The byte you want to add to the existing crc
  * @return The new CRC value
  */
-unsigned short crc16_arc_add(unsigned short crc, unsigned short byte);
+unsigned crc16_arc_add(unsigned crc, unsigned byte);
 
 /**
  * @brief Calculates the ARC CRC 16 of a given string
  * @param Data the string you want to turn into
  */
-unsigned short crc16_arc_compute(const char* data, size_t len);
+unsigned crc16_arc_compute(const char* data, size_t len);
